+++ core/src/se2/groupb/monopoly/screens/JoinGameScreen.java
@@ -50,7 +50,7 @@ public class JoinGameScreen implements Screen {
         buttonSizeY = (int) (Gdx.graphics.getHeight() / (4.545454 * 2));
 
         xPosButtons = (float) (Gdx.graphics.getWidth() / 2D - buttonSizeX / 2D);
-        yPosInitialButtons = Gdx.graphics.getHeight() / 20;
+        yPosInitialButtons = (float) (Gdx.graphics.getHeight() / 20D);
         yPosOffsetButtons = (float) (-Gdx.graphics.getWidth() / 8D);
 
         font = new BitmapFont();
@@ -93,7 +93,7 @@ public class JoinGameScreen implements Screen {
         if (buttonPressed) {
             // if server response: client connected is still missing
             font.draw(monopoly.batch, waitingText,
-                    (Gdx.graphics.getWidth() / 2 - waitingText.width / 2), (yPosInitialButtons + 1.5f * buttonSizeY));
+                    (float) (Gdx.graphics.getWidth() / 2D - waitingText.width / 2D), (yPosInitialButtons + 1.5f * buttonSizeY));
         }
 
         monopoly.batch.end();
